+++ src/layout/Header/Navigation/Navigation.module.scss
@@ -1,25 +1,26 @@
-@import '../../styles/vars';
+@import '../../../styles/vars';
 
 .navigation {
   display: grid;
   grid-template-columns: 100px 1fr 100px;
 
-  padding-top: 30px;
-  padding-bottom: 30px;
+  padding-top: 25px;
+  padding-bottom: 25px;
 }
 
 .logo {
+  grid-column: 1 / 2;
   place-self: center center;
 
   display: flex;
   justify-content: center;
   align-items: center;
 
-  padding: 8px;
-  transition: opacity 300ms;
+  padding: 15px;
+  transition: opacity $transitionDuration;
 
   &:hover {
-    opacity: 0.6;
+    opacity: $hoverOpacity;
   }
 }
 
@@ -29,7 +30,7 @@
 
   display: flex;
   justify-content: space-between;
-  gap: 80px;
+  gap: 60px;
   
   list-style: none;
 
@@ -39,10 +40,10 @@
 
     color: $fontDarkTheme;
     text-decoration: none;
-    transition: opacity 300ms;
+    transition: opacity $transitionDuration;
 
     &:hover {
-      opacity: 0.6;
+      opacity: $hoverOpacity;
     }
   }
 }
