+++ src/styles/_vars.scss
@@ -1,5 +1,9 @@
 $rootFontSize: 24px;
 
+$hoverOpacity: 0.6;
+
+$transitionDuration: 300ms;
+
 $containerMaxWidth: 1305px;
 $containerPaddingLeft: 40px;
 $containerPaddingRight: 40px;
